fix(agent-manager): align review chip sides and share formatter
