+++ packages/kilo-vscode/webview-ui/agent-manager/review-comments.ts
@@ -1,6 +1,8 @@
 import type { ReviewComment, WorktreeFileDiff } from "../src/types/messages"
+import { formatReviewCommentMarkdown, formatReviewCommentsMarkdown } from "../src/utils/review-comment-markdown"
 
 export type { ReviewComment }
+export { formatReviewCommentMarkdown, formatReviewCommentsMarkdown }
 
 function lineCount(text: string): number {
   if (text.length === 0) return 0
@@ -22,39 +24,6 @@ export function extractLines(content: string, start: number, end: number): strin
     .split("\n")
     .slice(start - 1, end)
     .join("\n")
-}
-
-function escapeMarkdownInline(value: string): string {
-  return value.replace(/([\\`*_\[\]{}()#+\-!|])/g, "\\$1")
-}
-
-function fenceFor(value: string): string {
-  const matches = value.match(/`+/g) ?? []
-  const longest = matches.reduce((max, item) => Math.max(max, item.length), 0)
-  return "`".repeat(Math.max(3, longest + 1))
-}
-
-function formatCodeBlock(value: string): string[] {
-  const fence = fenceFor(value)
-  return [fence, value, fence]
-}
-
-export function formatReviewCommentMarkdown(comment: ReviewComment): string {
-  const lines = [`**${escapeMarkdownInline(comment.file)}** (line ${comment.line}):`]
-  if (comment.selectedText) {
-    lines.push(...formatCodeBlock(comment.selectedText))
-  }
-  lines.push(comment.comment)
-  return lines.join("\n")
-}
-
-export function formatReviewCommentsMarkdown(comments: ReviewComment[]): string {
-  const lines = ["## Review Comments", ""]
-  for (const comment of comments) {
-    lines.push(formatReviewCommentMarkdown(comment))
-    lines.push("")
-  }
-  return lines.join("\n")
 }
 
 export function sanitizeReviewComments(comments: ReviewComment[], diffs: WorktreeFileDiff[]): ReviewComment[] {
